requirements.txt
Django==2.1
django-environ==0.4.5
django-extensions==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
factory-boy==2.11.1
psycopg2-binary==2.7.5
pytest==3.7.4
pytest-cov==2.5.1
pytest-django==3.4.2
pytest-sugar==0.9.1
whitenoise==4.0
ipython==6.5.0
